fixing builders and builderers

+++ Java/core/src/main/java/com/amazon/randomcutforest/store/LeafStore.java
@@ -50,7 +50,7 @@ public class LeafStore extends IndexManager implements ILeafStore {
         mass = new int[capacity];
     }
 
-    public LeafStore(int[] pointIndex, int[] parentIndex, int[] mass, int[] freeIndexes, short freeIndexPointer) {
+    public LeafStore(int[] pointIndex, int[] parentIndex, int[] mass, int[] freeIndexes, int freeIndexPointer) {
         super(freeIndexes, freeIndexPointer);
 
         checkNotNull(pointIndex, "pointIndex must not be null");
